2nd commit
This has some major changes in it. I couldn’t figure out how to replace
the current carousel with a new one, in the same area. They just kept
loading one after the other so I changed it to where they all display
at once for now until I can figure out how to do what I want it to. I
also wanted to display an h2 with the name of the photographer when the
carousel loaded. However, if I put it in a div, it displays as an image.

I also need to put the google contact form in. Thanks for looking at
this for me. I think I’d eventually get it all but an extra set of eyes
is really nice for a beginner :)